requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
transformers==4.35.2
torch>=2.7.1
pytesseract==0.3.10
Pillow>=11.0.0
PyMuPDF>=1.24.0,<1.25.0
sentence-transformers
google-generativeai
requests==2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
